initialized test executable for testing
did i mention this is for testing purposes which involves testing? itll test the program for me when i want to have it test functionality. its pretty cool, testing is very useful as it tests the capabilities of the library by doing various things to test itself and make sure the tests pass